Requirement ##840 : Verify Demographic Data of Incoming C-STOREs
This commit includes a working version of the image reconciliation.

- Added Description column in the ReconcileQueue table. This field contains the searchable description of the related study and images.
- Added SeriesDescription column in the ReconcileQueueUid table.
- Modified ReconcileStudyItemProcessor to actual reconcile the images. It uses UpdateDicomFileCommand and ReconcileDicomFileCommand to update the demographics in the images and move them to the correct storage location. It also creates a "StudyProcess" entry for the reconciled images.  If images are sent more than once, subsequent copies will be discarded.

Notes:

1. The ReconcileQueue Description field contains the searchable information of the objects in the queue. The key and value for each searchable field are separated by "=" and each field occupies a single line:

ExistingPatientsName=John^Smith
NewPatientsName=John^Doe

2. When user reconciles the images using the Web GUI (to be developed), the assumption is that the instructions will be stored in the ChangeDescription field of the StudyHistory record associated with the ReconcileStudy record. The instructions to update the images are stored in the "UpdateImages" node of the xml. For eg, to modify the patient's name in the images to John^Smith, the ChangeDescription should contain:

<UpdateImages>
    <SetTag  TagPath="00100010" Value="John^Smith"/>
</UpdateImages>

+++ ImageServer/Services/WorkQueue/ReconcileStudy/ReconcileStudyItemProcessor.cs
@@ -2,15 +2,22 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Text;
 using System.Xml;
 using ClearCanvas.Common;
+using ClearCanvas.Dicom;
+using ClearCanvas.Enterprise.Core;
 using ClearCanvas.ImageServer.Common;
+using ClearCanvas.ImageServer.Common.CommandProcessor;
 using ClearCanvas.ImageServer.Common.Utilities;
 using ClearCanvas.ImageServer.Model;
+using ClearCanvas.ImageServer.Model.Brokers;
+using ClearCanvas.ImageServer.Model.EntityBrokers;
+using ClearCanvas.ImageServer.Model.Parameters;
+using ClearCanvas.ImageServer.Services.WorkQueue.StudyProcess;
 
 namespace ClearCanvas.ImageServer.Services.WorkQueue.ReconcileStudy
 {
+
     /// <summary>
     /// Processor to handle 'ReconcileStudy' work queue entries
     /// </summary>
@@ -18,6 +25,9 @@ namespace ClearCanvas.ImageServer.Services.WorkQueue.ReconcileStudy
     {
         #region Private Members
         private ReconcileStudyProcessorContext _context;
+        private ReconcileDicomFileCommand _reconcileCommand;
+        private UpdateDicomFileCommand _updateDicomFileCommand;
+        private ReconcileStudyWorkQueueData _reconcileQueueData;
         #endregion
 
         #region Overridden Protected Method
@@ -25,20 +35,62 @@ namespace ClearCanvas.ImageServer.Services.WorkQueue.ReconcileStudy
         {
             Platform.CheckForNullReference(item, "item");
             Platform.CheckForNullReference(item.Data, "item.Data");
+            Platform.CheckForNullReference(item.StudyHistoryKey, "item.StudyHistoryKey");
 
-            LoadStorageLocation(item);
-            LoadUids(item);
+            if (CannotStart())
+            {
+                WorkQueueSettings settings = WorkQueueSettings.Instance;
+                PostponeItem(item, Platform.Time.Add(TimeSpan.FromMinutes(settings.WorkQueueQueryDelay)),
+                                  Platform.Time.Add(TimeSpan.FromSeconds(settings.WorkQueueExpireDelaySeconds)));
+            }
+            else
+            {
+                LoadUids(item);
 
-            SetupContext(item);
+                if (WorkQueueUidList.Count == 0)
+                {
+                    Complete(item);
+                }
+                else
+                {
+                    _reconcileQueueData = XmlUtils.Deserialize<ReconcileStudyWorkQueueData>(WorkQueueItem.Data);
 
-            ProcessUidList();
+                    LoadStorageLocation(item);
 
-            PostProcessing(item, true, WorkQueueUidList.Count==0);
+                    ProcessUidList();
+
+                    PostProcessing(item, WorkQueueUidList.Count > 0, WorkQueueUidList.Count == 0);
+                }
+            }
+           
+            
         }
+
+        
         #endregion
 
 
         #region Private Methods
+
+
+        private bool CannotStart()
+        {
+            // cannot start if the existing study is scheduled for update
+            WorkQueueSelectCriteria criteria = new WorkQueueSelectCriteria();
+            criteria.ServerPartitionKey.EqualTo(WorkQueueItem.ServerPartitionKey);
+            criteria.StudyStorageKey.EqualTo(WorkQueueItem.StudyStorageKey);
+            criteria.WorkQueueTypeEnum.In(new WorkQueueTypeEnum[] { WorkQueueTypeEnum.WebEditStudy });
+
+            List<Model.WorkQueue> items = FindRelatedWorkQueueItems(WorkQueueItem, criteria);
+            return items.Count > 0;
+        }
+
+
+        private void Complete(Model.WorkQueue item)
+        {
+            PostProcessing(item, false, true);
+        }
+
         private void ProcessUidList()
         {
             foreach(WorkQueueUid uid in WorkQueueUidList)
@@ -47,43 +99,87 @@ namespace ClearCanvas.ImageServer.Services.WorkQueue.ReconcileStudy
             }
         }
 
-        private void SetupContext(Model.WorkQueue item)
+        private void SetupCommands(DicomFile file)
         {
-            _context = new ReconcileStudyProcessorContext();
-            _context.WorkQueueItem = item;
-            _context.Partition = ServerPartition.Load(StorageLocation.ServerPartitionKey);
-            _context.Data = XmlUtils.Deserialize<ReconcileStudyWorkQueueData>(_context.WorkQueueItem.Data);
+            if (_updateDicomFileCommand==null)
+            {
+                XmlDocument changeDescXml = _context.History.ChangeDescription;
+                XmlNode updateImagesNode = changeDescXml.SelectSingleNode("UpdateImages");
+                if (updateImagesNode != null)
+                {
+                    UpdateImagesCommandXmlSpecification specs = new UpdateImagesCommandXmlSpecification();
+                    specs.Parse(changeDescXml.SelectSingleNode("UpdateImages"));
+                    _updateDicomFileCommand = new UpdateDicomFileCommand(specs.ActionList);
+                }
+            }
+
+            if (_reconcileCommand==null)
+                _reconcileCommand = new ReconcileDicomFileCommand();
+
+
+            _updateDicomFileCommand.DicomFile = file;
+            _reconcileCommand.Context = _context;
+            _reconcileCommand.DicomFile = file;
         }
+
+        private void SetupContext(DicomFile file)
+        {
+            Platform.CheckForNullReference(StorageLocation, "StorageLocation");
+
+            _context = new ReconcileStudyProcessorContext();
+            _context.WorkQueueItem = WorkQueueItem;
+            _context.Partition = ServerPartition.Load(StorageLocation.ServerPartitionKey);
+            _context.Data = _reconcileQueueData;
+            _context.History = StudyHistory.Load(WorkQueueItem.StudyHistoryKey);
+            _context.ExistingStudyStorageLocation = StorageLocation;   
+        }
+
 
         private void ReconcileUid(WorkQueueUid uid)
         {
             string imagePath = GetCurrentImagePath(uid);
             Debug.Assert(File.Exists(imagePath));
-
+            
             Platform.Log(LogLevel.Info, "Reconciling image : {0} in {1}", uid.SopInstanceUid, imagePath);
-
             ProcessFile(uid, imagePath);
 
             DeleteWorkQueueUid(uid);
-
+            
         }
 
         private string GetCurrentImagePath(WorkQueueUid uid)
         {
-            string path = Path.Combine(_context.Data.StoragePath, uid.SopInstanceUid +".dcm");
+            string path = Path.Combine(_reconcileQueueData.StoragePath, uid.SopInstanceUid +".dcm");
             return path;
         }
 
         private void ProcessFile(WorkQueueUid uid, string path)
         {
-            // TODO: Simulate what the StorageSCP does here
-            // create a StorageLocation if neeeded.
-            // create the folder for the StorageLocation if needed
-            // store the image files in this folder
-            // create StudyProcess work queue entry
+            
+            DicomFile file = new DicomFile(path);
+            file.Load(DicomReadOptions.StorePixelDataReferences);
 
-            File.Delete(path);
-            DirectoryUtility.DeleteIfEmpty(Directory.GetParent(path).FullName);
+            SetupContext(file);
+            SetupCommands(file);
+
+            using(ServerCommandProcessor processor = new ServerCommandProcessor("Reconcile File"))
+            {
+                processor.AddCommand(_updateDicomFileCommand);
+                processor.AddCommand(_reconcileCommand);
+
+                if (!processor.Execute())
+                {
+                    throw new ApplicationException(
+                        String.Format("Unable to reconcile dicom file {0} : {1}", path, processor.FailureReason));
+                }
+                else
+                {
+                    // image has been reconciled. It can be deleted.
+                    File.Delete(path);
+                    DirectoryUtility.DeleteIfEmpty(Directory.GetParent(path).FullName);
+                }
+            }
+            
         }
 
         #endregion
